Uploaded new UML
I created a new UML with a twitter model and in the source program (even though I am unable to find it in the git).  The problem is that lucidchart has a limit of things I can put in without going over the limit, so I had the combine the Instagram, Posts, Youtube, and Twitter templates into one chart.